adapted example for typescript
