+++ src/index.tsx
@@ -1,40 +1,45 @@
 import React, { Props, ReactElement, ComponentType, useEffect, useState } from 'react'
 
-export type Dispatcher = (action: IAction) => void
 export interface IAction {
   type: string;
   payload: any;
   [key: string]: any;
 }
+export type Dispatcher = (action: IAction) => void
 export interface IBaseState {
   dispatch: Dispatcher;
 }
 
 export default function<State>(
   initialState: State,
+  reducers: Array<(state: State, action: IAction, dispatcher?: Dispatcher) => State>,
   middlewares: Array<{
-    before: (state: State, action: IAction, dispatcher?: Dispatcher) => void,
-    after: (state: State, action: IAction, dispatcher?: Dispatcher) => void,
+    before?: (state: State, action: IAction, dispatcher?: Dispatcher) => void,
+    after?: (state: State, action: IAction, dispatcher?: Dispatcher) => void,
   }> = [],
 ) {
-  
+  let dispatcher: Dispatcher = (action) => {
+    actionStack.push(action)
+  }
+
   type Reducer = (state: State, action: IAction, dispatcher?: Dispatcher) => State
   type Selector = (state: State, props?: Props<any>) => Partial<State>;
 
   const actionStack: Array<IAction> = []
-  let reducer: Reducer
-  let dispatch: Dispatcher = (action) => actionStack.push(action)
-
-  const combineReducers = (reducers: Reducer[]) => {
-    reducer = (state, action, dispatch) => reducers.reduce((newState, r) => r(newState, action, dispatch), state)
-  }
-
+  
+  const reducer: Reducer = (state, action, dispatch) => (
+    reducers.reduce((newState, r) => r(newState, action, dispatch), state)
+  )
+  
+  const dispatch: Dispatcher = (action) => dispatcher(action)
+  
   const Context = React.createContext<State>({
     ...initialState,
     dispatch: (action: IAction) => {
       console.warn('Please add a Provider to your application first')
     }
   })
+  
   const Consumer = Context.Consumer
 
   const useSelector = (select: Selector) => select(React.useContext(Context))
@@ -56,19 +61,16 @@ export default function<State>(
 
   const Provider: React.FC<{ children: ReactElement }> = ({ children }) => {
 
-    const [state, setState] = useState<State>(initialState)
+    const [state, setState] = useState<State>({ ...initialState, dispatch })
     
-    dispatch = (action: IAction) => setState((state) => {
+    dispatcher = (action: IAction) => setState((state) => {
       runMiddlewares('before', state, action, dispatch)
       const newState = reducer(state, action, dispatch)
       runMiddlewares('after', newState, action, dispatch)
       return newState
     })
     
-    useEffect(() => {
-      setState((state) => ({ ...state, dispatch }))
-      actionStack.forEach(dispatch)
-    }, [])
+    useEffect(() => actionStack.forEach(dispatch), [])
     
     return (
       <Context.Provider value={state}>
@@ -76,5 +78,5 @@ export default function<State>(
       </Context.Provider>
     )
   }
-  return { Provider, withState, useSelector, dispatch, combineReducers }
+  return { Provider, withState, useSelector, dispatch }
 }
